Add authorisation to endpoints

+++ src/main/java/org/kainos/ea/resources/DeliveryEmployeeController.java
@@ -2,7 +2,9 @@ package org.kainos.ea.resources;
 
 import io.swagger.annotations.Api;
 import org.kainos.ea.api.DeliveryEmployeeService;
+import org.kainos.ea.auth.AccessValidator;
 import org.kainos.ea.cli.DeliveryEmployeeRequest;
+import org.kainos.ea.cli.Role;
 import org.kainos.ea.client.*;
 import org.kainos.ea.core.DeliveryEmployeeValidator;
 import org.kainos.ea.db.DeliveryEmployeeDao;
@@ -23,19 +25,27 @@ public class DeliveryEmployeeController {
     @GET
     @Path("/deliveryemployee")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getDeliveryEmployee() {
+    public Response getDeliveryEmployee(@QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             return Response.ok(deliveryEmployeeService.getDeliveryEmployees()).build();
         } catch (FailedToGetDeliveryEmployeesException e) {
             return Response.serverError().build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
     }
 
     @GET
     @Path("/deliveryemployee/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getDeliveryEmployeeById(@PathParam("id") int id) {
+    public Response getDeliveryEmployeeById(@PathParam("id") int id, @QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             return Response.ok(deliveryEmployeeService.getDeliveryEmployeeById(id)).build();
         } catch (FailedToGetDeliveryEmployeesException e) {
             return Response.serverError().build();
@@ -43,14 +53,20 @@ public class DeliveryEmployeeController {
             System.err.println(e.getMessage());
 
             return Response.status(Response.Status.BAD_REQUEST).build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
     }
 
     @POST
     @Path("/deliveryemployee")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response createDeliveryEmployee(DeliveryEmployeeRequest request) {
+    public Response createDeliveryEmployee(DeliveryEmployeeRequest request, @QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             return Response.ok(deliveryEmployeeService.createDeliveryEmployee(request)).build();
         } catch (FailedToCreateDeliveryEmployeeException e) {
             e.printStackTrace();
@@ -60,14 +76,20 @@ public class DeliveryEmployeeController {
             e.printStackTrace();
 
             return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
     }
 
     @PUT
     @Path("/deliveryemployee/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response updateDeliveryEmployee(@PathParam("id") int id, DeliveryEmployeeRequest request) {
+    public Response updateDeliveryEmployee(@PathParam("id") int id, DeliveryEmployeeRequest request, @QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             deliveryEmployeeService.updateDeliveryEmployee(id, request);
 
             return Response.ok().build();
@@ -79,14 +101,20 @@ public class DeliveryEmployeeController {
             e.printStackTrace();
 
             return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
     }
 
     @DELETE
     @Path("/deliveryemployee/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response deleteDeliveryEmployeeId(@PathParam("id") int id) {
+    public Response deleteDeliveryEmployeeId(@PathParam("id") int id, @QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             deliveryEmployeeService.deleteDeliveryEmployee(id);
 
             return Response.ok().build();
@@ -98,6 +126,10 @@ public class DeliveryEmployeeController {
             e.printStackTrace();
 
             return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
     }
 }
